requirements.txt
fastapi[standard]
pydantic
psycopg[binary]
python-dotenv
bcrypt
pyjwt
argparse

opencv-python
numpy
pillow
